backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
clickhouse-connect==0.8.17
comm==0.2.2
confluent-kafka==2.10.0
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.56.0
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
email-validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
faker==37.4.0
fastapi==0.115.12
fastapi-cli==0.0.7
fonttools==4.58.2
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.8
lz4==4.4.4
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
prompt-toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py4j==0.10.9.9
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.5
pydantic-core==2.33.2
pygments==2.19.1
pyparsing==3.2.3
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
tornado==6.5.1
traitlets==5.14.3
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
zstandard==0.23.0
